# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthErrorResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "AuthTokensResponse": {
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "accessTokenExpiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ChangePasswordRequest": {
        "properties": {
          "currentPassword": {
            "minLength": 1,
            "type": "string"
          },
          "newPassword": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "cookieMode": {
            "type": "boolean"
          },
          "deviceName": {
            "type": "string"
          },
          "email": {
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
      "ReauthRequest": {
        "properties": {
          "password": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "password"
        ],
        "type": "object"
      },
...
      "RefreshRequest": {
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "SelectHouseholdRequest": {
        "properties": {
          "householdId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "householdId"
        ],
        "type": "object"
      },
      "SelectProfileRequest": {
        "properties": {
          "pin": {
            "type": "string"
          },
          "profileId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "profileId"
        ],
        "type": "object"
      },
      "SetupRequest": {
        "properties": {
          "cookieMode": {
            "type": "boolean"
          },
          "displayName": {
            "minLength": 1,
            "type": "string"
          },
          "email": {
            "minLength": 1,
            "type": "string"
          },
          "householdName": {
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "minLength": 1,
            "type": "string"
          },
          "profileName": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "displayName",
          "email",
          "householdName",
          "password",
          "profileName"
        ],
        "type": "object"
      },
      "StatusResponse": {
        "properties": {
          "devicePairingEnabled": {
            "type": "boolean"
          },
          "setupComplete": {
            "type": "boolean"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "accessCookie": {
        "description": "JWT access token. Profile tokens also satisfy SCOPE_ACCOUNT requirements. Unsafe cookie-authenticated requests also require the X-XSRF-TOKEN header.",
        "in": "cookie",
        "name": "streamarr_access",
        "type": "apiKey"
      },
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "JWT access token. Profile tokens also satisfy SCOPE_ACCOUNT requirements.",
        "scheme": "bearer",
        "type": "http"
      },
      "playbackQuery": {
        "description": "Playback token bound to this stream session. Stream endpoints ignore bearer headers and access cookies.",
        "in": "query",
        "name": "t",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Streamarr",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/change-password": {
      "post": {
        "operationId": "changePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Refusal: route on `code`, `message` is displayable"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "SCOPE_ACCOUNT"
            ]
          },
          {
            "accessCookie": [
              "SCOPE_ACCOUNT"
            ]
          }
        ],
        "tags": [
          "auth-controller"
        ]
      }
    },
...
    "/api/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Refusal: route on `code`, `message` is displayable"
          }
        },
        "security": [ ],
        "tags": [
          "auth-controller"
        ]
      }
    },
...
    "/api/auth/reauth": {
      "post": {
        "operationId": "reauthenticate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReauthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Refusal: route on `code`, `message` is displayable"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "SCOPE_ACCOUNT"
            ]
          },
          {
            "accessCookie": [
              "SCOPE_ACCOUNT"
            ]
          }
        ],
        "tags": [
          "auth-controller"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Refusal: route on `code`, `message` is displayable"
          }
        },
        "security": [ ],
        "tags": [
          "auth-controller"
        ]
      }
    },
    "/api/auth/refresh/revoke": {
      "post": {
        "operationId": "logout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Refusal: route on `code`, `message` is displayable"
          }
        },
        "security": [ ],
        "tags": [
          "auth-controller"
        ]
      }
    },
    "/api/auth/select-household": {
      "post": {
        "operationId": "selectHousehold",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectHouseholdRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Refusal: route on `code`, `message` is displayable"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "SCOPE_ACCOUNT"
            ]
          },
          {
            "accessCookie": [
              "SCOPE_ACCOUNT"
            ]
          }
        ],
        "tags": [
          "auth-controller"
        ]
      }
    },
    "/api/auth/select-profile": {
      "post": {
        "operationId": "selectProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Refusal: route on `code`, `message` is displayable"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "SCOPE_ACCOUNT"
            ]
          },
          {
            "accessCookie": [
              "SCOPE_ACCOUNT"
            ]
          }
        ],
        "tags": [
          "auth-controller"
        ]
      }
    },
    "/api/auth/setup": {
      "post": {
        "operationId": "setup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensResponse"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Refusal: route on `code`, `message` is displayable"
          }
        },
        "security": [ ],
        "tags": [
          "auth-controller"
        ]
      }
    },
    "/api/auth/status": {
      "get": {
        "operationId": "status",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Refusal: route on `code`, `message` is displayable"
          }
        },
        "security": [ ],
        "tags": [
          "auth-controller"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}